+++ engine/maths/primes.h
@@ -84,7 +84,7 @@ namespace regina {
  */
 class Primes {
     private:
-        static const unsigned long numPrimeSeeds;
+        static const size_t numPrimeSeeds;
             /**< The size of the hard-coded list of seed primes. */
         static const unsigned long primeSeedList[];
             /**< The full hard-coded list of seed primes. */
@@ -112,7 +112,7 @@ class Primes {
          *
          * \return the number of primes or suspected primes currently stored.
          */
-        static unsigned long size();
+        static size_t size();
 
         /**
          * Returns the requested prime (or suspected prime).  More
@@ -141,7 +141,7 @@ class Primes {
          * \return the requested prime (or suspected prime), or zero if
          * \a which was too large and \a autoGrow was \c false.
          */
-        static Integer prime(unsigned long which, bool autoGrow = true);
+        static Integer prime(size_t which, bool autoGrow = true);
 
         /**
          * Returns the prime factorisation of the given integer as a list
@@ -257,12 +257,12 @@ class Primes {
          * \param extras the number of additional suspected primes to
          * calculate.
          */
-        static void growPrimeList(unsigned long extras = 1);
+        static void growPrimeList(size_t extras = 1);
 };
 
 // Inline functions for Primes
 
-inline unsigned long Primes::size() {
+inline size_t Primes::size() {
     std::lock_guard<std::mutex> lock(largeMutex);
     return numPrimeSeeds + largePrimes.size();
 }
