+++ engine/maths/seedprimes.cpp
@@ -37,7 +37,7 @@ namespace regina {
 // Keep the seed primes in a separate file so they're nicely out of the
 // way of everything else.
 
-const unsigned long Primes::numPrimeSeeds = 10000;
+const size_t Primes::numPrimeSeeds = 10000;
 const unsigned long Primes::primeSeedList[] = {
       2,      3,      5,      7,     11,     13,     17,     19,     23,     29,
      31,     37,     41,     43,     47,     53,     59,     61,     67,     71,
